Refactor source into additional headers and c files
To support use of the VHD and VHDX classes for other uses besides
the ConvertImage function.
